refactor: Extract logic for fetching location to the details view model
