+++ app/src/main/java/uni/aeh/tasktracker/details/ui/DetailsScreen.kt
@@ -1,10 +1,6 @@
 package uni.aeh.tasktracker.details.ui
 
-import android.Manifest
 import android.content.Context
-import android.content.pm.PackageManager
-import android.location.Location
-import android.util.Log
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
@@ -12,20 +8,15 @@ import androidx.compose.material3.Button
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
+import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.core.app.ActivityCompat
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.navigation.NavController
 import androidx.navigation.compose.rememberNavController
-import com.google.android.gms.location.LocationRequest
-import com.google.android.gms.location.LocationServices
-import com.google.android.gms.tasks.CancellationTokenSource
 import kotlinx.coroutines.launch
 import uni.aeh.tasktracker.Screen
 
@@ -38,7 +29,7 @@ fun DetailsScreen(navController: NavController, context: Context = LocalContext.
     // or active, preventing memory leaks and ensuring proper resource management.
     val coroutineScope = rememberCoroutineScope()
 
-    val (location, setLocation) = remember { mutableStateOf<Location?>(null) }
+    val location = viewModel.location.collectAsState()
 
     LaunchedEffect(Unit) {
         coroutineScope.launch {
@@ -57,18 +48,19 @@ fun DetailsScreen(navController: NavController, context: Context = LocalContext.
     ) {
         Button(
             onClick = {
-                getCurrentLocation(context, setLocation)
+                viewModel.getCurrentLocation(context)
             }
         ) {
             Text(text = "Get Current Location")
         }
 
-        if (location != null) {
+        location.value?.let { location ->
             Text(text = "Latitude: ${location.latitude}")
             Text(text = "Longitude: ${location.longitude}")
-        } else {
+        } ?: run {
             Text(text = "You need to grant location permissions in order for this feature to work correctly")
         }
+
 
         Button(onClick = viewModel::buttonClicked) {
             Text(text = "details effect")
@@ -86,41 +78,4 @@ private fun DetailsScreenPreview() {
     val navController = rememberNavController()
 
     DetailsScreen(navController)
-}
-
-
-private fun getCurrentLocation(context: Context, setLocation: (Location?) -> Unit) {
-    val fusedLocationProviderClient = LocationServices.getFusedLocationProviderClient(context)
-
-    val locationRequest = LocationRequest.Builder(1000L)
-        .setMinUpdateDistanceMeters(100f)
-        .build()
-    val cancellationTokenSource = CancellationTokenSource()
-
-    if (ActivityCompat.checkSelfPermission(
-            context,
-            Manifest.permission.ACCESS_FINE_LOCATION
-        ) != PackageManager.PERMISSION_GRANTED && ActivityCompat.checkSelfPermission(
-            context,
-            Manifest.permission.ACCESS_COARSE_LOCATION
-        ) != PackageManager.PERMISSION_GRANTED
-    ) {
-        Log.d("Location", "Permission not granted")
-        return
-    }
-
-    fusedLocationProviderClient.getCurrentLocation(
-        locationRequest.priority,
-        cancellationTokenSource.token
-    ).addOnSuccessListener { location: Location ->
-        Log.d("Location", "Location: $location")
-        val latitude = location.latitude
-        val longitude = location.longitude
-        // Do something with the latitude and longitude
-        Log.d("Location", "Latitude: $latitude, Longitude: $longitude")
-        setLocation(location)
-
-    }.addOnFailureListener { exception: Exception ->
-        Log.d("Location", "error: $exception")
-    }
 }
